+++ drivers/net/ethernet/intel/ixgbe/ixgbe_ipsec.c
@@ -775,7 +775,7 @@ static void ixgbe_ipsec_del_sa(struct net_device *dev, struct xfrm_state *xs)
 
 		if (!rsa->used) {
 			netdev_err(dev, "Invalid Rx SA selected sa_idx=%d offload_handle=%lu\n",
-				   sa_idx, xs->xso.offload_handle);
+				   sa_idx, (unsigned long)xs->xso.offload_handle);
 			return;
 		}
 
@@ -804,7 +804,7 @@ static void ixgbe_ipsec_del_sa(struct net_device *dev, struct xfrm_state *xs)
 
 		if (!ipsec->tx_tbl[sa_idx].used) {
 			netdev_err(dev, "Invalid Tx SA selected sa_idx=%d offload_handle=%lu\n",
-				   sa_idx, xs->xso.offload_handle);
+				   sa_idx, (unsigned long)xs->xso.offload_handle);
 			return;
 		}
 
@@ -1080,7 +1080,7 @@ int ixgbe_ipsec_tx(struct ixgbe_ring *tx_ring,
 	itd->sa_idx = xs->xso.offload_handle - IXGBE_IPSEC_BASE_TX_INDEX;
 	if (unlikely(itd->sa_idx >= IXGBE_IPSEC_MAX_SA_COUNT)) {
 		netdev_err(tx_ring->netdev, "%s: bad sa_idx=%d handle=%lu\n",
-			   __func__, itd->sa_idx, xs->xso.offload_handle);
+			   __func__, itd->sa_idx, (unsigned long)xs->xso.offload_handle);
 		return 0;
 	}
