+++ drivers/net/ethernet/mellanox/mlx5/core/en_tc.c
@@ -4841,7 +4841,7 @@ int mlx5e_configure_flower(struct net_device *dev, struct mlx5e_priv *priv,
 				   "flow cookie already exists, ignoring");
 		netdev_warn_once(priv->netdev,
 				 "flow cookie %lx already exists, ignoring\n",
-				 f->cookie);
+				 (unsigned long)f->cookie);
 		err = -EEXIST;
 		goto rcu_unlock;
 	}
